+++ tests/common.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+TEST_AGENT = 'narwal (goo.gl/IBenG) testing'