+++ HISTORY.rst
@@ -3,6 +3,16 @@
 Changelog
 ---------
 
+Development
++++++++++++
+* added limit parameter to Reddit.info()
+* changed Reddit._subreddit_get() parameter order and removed default values
+* fixed default value for items in ListBlob to be list() instead of []
+* added list methods to ListBlob
+* fixed Thing.__unicode__() to always return a string (u'' instead of None)
+* added utf-8 coding declaration to source files
+
+
 v0.2.4a (2012-04-23)
 ++++++++++++++++++++
 * automatically un-html-escape unicode chars